App/requirements.txt
tensorflow-lite
numpy
scikit-learn
skops